Moved indexing routines for host reference code into a unified file; began working towards eliminating separate single and multiple GPU verification code
