+++ tests/host_reference/domain_wall_dslash_reference.cpp
@@ -13,217 +13,10 @@
 #include <gauge_field.h>
 #include <color_spinor_field.h>
 
+#include <gamma_reference.h>
+
 using namespace quda;
 
-// i represents a "half index" into an even or odd "half lattice".
-// when oddBit={0,1} the half lattice is {even,odd}.
-//
-// the displacements, such as dx, refer to the full lattice coordinates.
-//
-// neighborIndex() takes a "half index", displaces it, and returns the
-// new "half index", which can be an index into either the even or odd lattices.
-// displacements of magnitude one always interchange odd and even lattices.
-//
-//
-int neighborIndex_4d(int i, int oddBit, int dx4, int dx3, int dx2, int dx1)
-{
-  // On input i should be in the range [0 , ... , Z[0]*Z[1]*Z[2]*Z[3]/2-1].
-  if (i < 0 || i >= (Z[0] * Z[1] * Z[2] * Z[3] / 2)) {
-    printf("i out of range in neighborIndex_4d\n");
-    exit(-1);
-  }
-  // Compute the linear index.  Then dissect.
-  // fullLatticeIndex_4d is in util_quda.cpp.
-  // The gauge fields live on a 4d sublattice.
-  int X = fullLatticeIndex_4d(i, oddBit);
-  int x4 = X / (Z[2] * Z[1] * Z[0]);
-  int x3 = (X / (Z[1] * Z[0])) % Z[2];
-  int x2 = (X / Z[0]) % Z[1];
-  int x1 = X % Z[0];
-
-  x4 = (x4 + dx4 + Z[3]) % Z[3];
-  x3 = (x3 + dx3 + Z[2]) % Z[2];
-  x2 = (x2 + dx2 + Z[1]) % Z[1];
-  x1 = (x1 + dx1 + Z[0]) % Z[0];
-
-  return (x4 * (Z[2] * Z[1] * Z[0]) + x3 * (Z[1] * Z[0]) + x2 * (Z[0]) + x1) / 2;
-}
-
-//#ifndef MULTI_GPU
-// This is just a copy of gaugeLink() from the quda code, except
-// that neighborIndex() is replaced by the renamed version
-// neighborIndex_4d().
-// ok
-template <typename Float> Float *gaugeLink_sgpu(int i, int dir, int oddBit, Float **gaugeEven, Float **gaugeOdd)
-{
-  Float **gaugeField;
-  int j;
-
-  // If going forward, just grab link at site, U_\mu(x).
-  if (dir % 2 == 0) {
-    j = i;
-    // j will get used in the return statement below.
-    gaugeField = (oddBit ? gaugeOdd : gaugeEven);
-  } else {
-    // If going backward, a shift must occur, U_\mu(x-\muhat)^\dagger;
-    // dagger happens elsewhere, here we're just doing index gymnastics.
-    switch (dir) {
-    case 1: j = neighborIndex_4d(i, oddBit, 0, 0, 0, -1); break;
-    case 3: j = neighborIndex_4d(i, oddBit, 0, 0, -1, 0); break;
-    case 5: j = neighborIndex_4d(i, oddBit, 0, -1, 0, 0); break;
-    case 7: j = neighborIndex_4d(i, oddBit, -1, 0, 0, 0); break;
-    default: j = -1; break;
-    }
-    gaugeField = (oddBit ? gaugeEven : gaugeOdd);
-  }
-
-  return &gaugeField[dir / 2][j * (3 * 3 * 2)];
-}
-
-//#else
-
-// Standard 4d version (nothing to change)
-template <typename Float>
-Float *gaugeLink_mgpu(int i, int dir, int oddBit, Float **gaugeEven, Float **gaugeOdd, Float **ghostGaugeEven,
-                      Float **ghostGaugeOdd, int n_ghost_faces, int nbr_distance)
-{
-  Float **gaugeField;
-  int j;
-  int d = nbr_distance;
-  if (dir % 2 == 0) {
-    j = i;
-    gaugeField = (oddBit ? gaugeOdd : gaugeEven);
-  } else {
-
-    int Y = fullLatticeIndex(i, oddBit);
-    int x4 = Y / (Z[2] * Z[1] * Z[0]);
-    int x3 = (Y / (Z[1] * Z[0])) % Z[2];
-    int x2 = (Y / Z[0]) % Z[1];
-    int x1 = Y % Z[0];
-    int X1 = Z[0];
-    int X2 = Z[1];
-    int X3 = Z[2];
-    int X4 = Z[3];
-    Float *ghostGaugeField;
-
-    switch (dir) {
-    case 1: { //-X direction
-      int new_x1 = (x1 - d + X1) % X1;
-      if (x1 - d < 0 && comm_dim_partitioned(0)) {
-        ghostGaugeField = (oddBit ? ghostGaugeEven[0] : ghostGaugeOdd[0]);
-        int offset = (n_ghost_faces + x1 - d) * X4 * X3 * X2 / 2 + (x4 * X3 * X2 + x3 * X2 + x2) / 2;
-        return &ghostGaugeField[offset * (3 * 3 * 2)];
-      }
-      j = (x4 * X3 * X2 * X1 + x3 * X2 * X1 + x2 * X1 + new_x1) / 2;
-      break;
-    }
-    case 3: { //-Y direction
-      int new_x2 = (x2 - d + X2) % X2;
-      if (x2 - d < 0 && comm_dim_partitioned(1)) {
-        ghostGaugeField = (oddBit ? ghostGaugeEven[1] : ghostGaugeOdd[1]);
-        int offset = (n_ghost_faces + x2 - d) * X4 * X3 * X1 / 2 + (x4 * X3 * X1 + x3 * X1 + x1) / 2;
-        return &ghostGaugeField[offset * (3 * 3 * 2)];
-      }
-      j = (x4 * X3 * X2 * X1 + x3 * X2 * X1 + new_x2 * X1 + x1) / 2;
-      break;
-    }
-    case 5: { //-Z direction
-      int new_x3 = (x3 - d + X3) % X3;
-      if (x3 - d < 0 && comm_dim_partitioned(2)) {
-        ghostGaugeField = (oddBit ? ghostGaugeEven[2] : ghostGaugeOdd[2]);
-        int offset = (n_ghost_faces + x3 - d) * X4 * X2 * X1 / 2 + (x4 * X2 * X1 + x2 * X1 + x1) / 2;
-        return &ghostGaugeField[offset * (3 * 3 * 2)];
-      }
-      j = (x4 * X3 * X2 * X1 + new_x3 * X2 * X1 + x2 * X1 + x1) / 2;
-      break;
-    }
-    case 7: { //-T direction
-      int new_x4 = (x4 - d + X4) % X4;
-      if (x4 - d < 0 && comm_dim_partitioned(3)) {
-        ghostGaugeField = (oddBit ? ghostGaugeEven[3] : ghostGaugeOdd[3]);
-        int offset = (n_ghost_faces + x4 - d) * X1 * X2 * X3 / 2 + (x3 * X2 * X1 + x2 * X1 + x1) / 2;
-        return &ghostGaugeField[offset * (3 * 3 * 2)];
-      }
-      j = (new_x4 * (X3 * X2 * X1) + x3 * (X2 * X1) + x2 * (X1) + x1) / 2;
-      break;
-    } // 7
-
-    default:
-      j = -1;
-      printf("ERROR: wrong dir \n");
-      exit(1);
-    }
-    gaugeField = (oddBit ? gaugeEven : gaugeOdd);
-  }
-
-  return &gaugeField[dir / 2][j * (3 * 3 * 2)];
-}
-
-// J  Directions 0..7 were used in the 4d code.
-// J  Directions 8,9 will be for P_- and P_+, chiral
-// J  projectors.
-const double projector[10][4][4][2] = {{{{1, 0}, {0, 0}, {0, 0}, {0, -1}},
-                                        {{0, 0}, {1, 0}, {0, -1}, {0, 0}},
-                                        {{0, 0}, {0, 1}, {1, 0}, {0, 0}},
-                                        {{0, 1}, {0, 0}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {0, 0}, {0, 1}},
-                                        {{0, 0}, {1, 0}, {0, 1}, {0, 0}},
-                                        {{0, 0}, {0, -1}, {1, 0}, {0, 0}},
-                                        {{0, -1}, {0, 0}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {0, 0}, {1, 0}},
-                                        {{0, 0}, {1, 0}, {-1, 0}, {0, 0}},
-                                        {{0, 0}, {-1, 0}, {1, 0}, {0, 0}},
-                                        {{1, 0}, {0, 0}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {0, 0}, {-1, 0}},
-                                        {{0, 0}, {1, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {1, 0}, {0, 0}},
-                                        {{-1, 0}, {0, 0}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {0, -1}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {0, 0}, {0, 1}},
-                                        {{0, 1}, {0, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {0, -1}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {0, 1}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {0, 0}, {0, -1}},
-                                        {{0, -1}, {0, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {0, 1}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {-1, 0}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {0, 0}, {-1, 0}},
-                                        {{-1, 0}, {0, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {-1, 0}, {0, 0}, {1, 0}}},
-                                       {{{1, 0}, {0, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {0, 0}, {1, 0}},
-                                        {{1, 0}, {0, 0}, {1, 0}, {0, 0}},
-                                        {{0, 0}, {1, 0}, {0, 0}, {1, 0}}},
-                                       // P_+ = P_R
-                                       {{{0, 0}, {0, 0}, {0, 0}, {0, 0}},
-                                        {{0, 0}, {0, 0}, {0, 0}, {0, 0}},
-                                        {{0, 0}, {0, 0}, {2, 0}, {0, 0}},
-                                        {{0, 0}, {0, 0}, {0, 0}, {2, 0}}},
-                                       // P_- = P_L
-                                       {{{2, 0}, {0, 0}, {0, 0}, {0, 0}},
-                                        {{0, 0}, {2, 0}, {0, 0}, {0, 0}},
-                                        {{0, 0}, {0, 0}, {0, 0}, {0, 0}},
-                                        {{0, 0}, {0, 0}, {0, 0}, {0, 0}}}};
-
-// todo pass projector
-template <typename Float> void multiplySpinorByDiracProjector5(Float *res, int projIdx, Float *spinorIn)
-{
-  for (int i = 0; i < 4 * 3 * 2; i++) res[i] = 0.0;
-
-  for (int s = 0; s < 4; s++) {
-    for (int t = 0; t < 4; t++) {
-      Float projRe = projector[projIdx][s][t][0];
-      Float projIm = projector[projIdx][s][t][1];
-
-      for (int m = 0; m < 3; m++) {
-        Float spinorRe = spinorIn[t * (3 * 2) + m * (2) + 0];
-        Float spinorIm = spinorIn[t * (3 * 2) + m * (2) + 1];
-        res[s * (3 * 2) + m * (2) + 0] += projRe * spinorRe - projIm * spinorIm;
-        res[s * (3 * 2) + m * (2) + 1] += projRe * spinorIm + projIm * spinorRe;
-      }
-    }
-  }
-}
 
 //#ifndef MULTI_GPU
 // dslashReference_4d()
@@ -268,14 +61,14 @@ void dslashReference_4d_sgpu(sFloat *res, gFloat **gaugeFull, sFloat *spinorFiel
         // xs=1.  Then the odd spinor site x1=x2=x3=x4=0 wants the even gauge array
         // element 0, so that we get U_\mu(0).
         gaugeOddBit = (xs % 2 == 0 || type == QUDA_4D_PC) ? oddBit : (oddBit + 1) % 2;
-        gFloat *gauge = gaugeLink_sgpu(gge_idx, dir, gaugeOddBit, gaugeEven, gaugeOdd);
+        gFloat *gauge = gaugeLink(gge_idx, dir, gaugeOddBit, gaugeEven, gaugeOdd, 1);
 
         // Even though we're doing the 4d part of the dslash, we need
         // to use a 5d neighbor function, to get the offsets right.
         sFloat *spinor = spinorNeighbor_5d<type>(sp_idx, dir, oddBit, spinorField);
         sFloat projectedSpinor[4 * 3 * 2], gaugedSpinor[4 * 3 * 2];
         int projIdx = 2 * (dir / 2) + (dir + daggerBit) % 2;
-        multiplySpinorByDiracProjector5(projectedSpinor, projIdx, spinor);
+        multiplySpinorByDiracProjector(projectedSpinor, projIdx, spinor);
 
         for (int s = 0; s < 4; s++) {
           if (dir % 2 == 0) {
@@ -320,13 +113,12 @@ void dslashReference_4d_mgpu(sFloat *res, gFloat **gaugeFull, gFloat **ghostGaug
       for (int dir = 0; dir < 8; dir++) {
         int gaugeOddBit = (xs % 2 == 0 || type == QUDA_4D_PC) ? oddBit : (oddBit + 1) % 2;
 
-        gFloat *gauge = gaugeLink_mgpu(i, dir, gaugeOddBit, gaugeEven, gaugeOdd, ghostGaugeEven, ghostGaugeOdd, 1,
-                                       1); // this is unchanged from MPi version
-        sFloat *spinor = spinorNeighbor_5d_mgpu<type>(sp_idx, dir, oddBit, spinorField, fwdSpinor, backSpinor, 1, 1);
+        gFloat *gauge = gaugeLink(i, dir, gaugeOddBit, gaugeEven, gaugeOdd, ghostGaugeEven, ghostGaugeOdd, 1, 1);
+        sFloat *spinor = spinorNeighbor_5d<type>(sp_idx, dir, oddBit, spinorField, fwdSpinor, backSpinor, 1, 1);
 
         sFloat projectedSpinor[spinor_site_size], gaugedSpinor[spinor_site_size];
         int projIdx = 2 * (dir / 2) + (dir + daggerBit) % 2;
-        multiplySpinorByDiracProjector5(projectedSpinor, projIdx, spinor);
+        multiplySpinorByDiracProjector(projectedSpinor, projIdx, spinor);
 
         for (int s = 0; s < 4; s++) {
           if (dir % 2 == 0)
@@ -388,7 +180,7 @@ void mdw_eofa_m5_ref(sFloat *res, sFloat *spinorField, int oddBit, int daggerBit
       sFloat *spinor = spinorNeighbor_5d<QUDA_4D_PC>(i, dir, oddBit, spinorField);
       sFloat projectedSpinor[spinor_size];
       int projIdx = 2 * (dir / 2) + (dir + daggerBit) % 2;
-      multiplySpinorByDiracProjector5(projectedSpinor, projIdx, spinor);
+      multiplySpinorByDiracProjector(projectedSpinor, projIdx, spinor);
       // J  Need a conditional here for s=0 and s=Ls-1.
       int X = fullLatticeIndex_5d_4dpc(i, oddBit);
       int xs = X / (Z[3] * Z[2] * Z[1] * Z[0]);
@@ -464,7 +256,7 @@ void dslashReference_5th(sFloat *res, sFloat *spinorField, int oddBit, int dagge
       sFloat *spinor = spinorNeighbor_5d<type>(i, dir, oddBit, spinorField);
       sFloat projectedSpinor[4 * 3 * 2];
       int projIdx = 2 * (dir / 2) + (dir + daggerBit) % 2;
-      multiplySpinorByDiracProjector5(projectedSpinor, projIdx, spinor);
+      multiplySpinorByDiracProjector(projectedSpinor, projIdx, spinor);
       // J  Need a conditional here for s=0 and s=Ls-1.
       int X = (type == QUDA_5D_PC) ? fullLatticeIndex_5d(i, oddBit) : fullLatticeIndex_5d_4dpc(i, oddBit);
       int xs = X / (Z[3] * Z[2] * Z[1] * Z[0]);
